Refactor out builtin to sygus.
